dividing 'templates' logic into two: library and testContracts
